Refactor prerelease registry functions folder and extract shared logic
